refactor: move getSelection from utils
